[Typesense]  Sync server state in getOrCreateCollectionFromModel #845 (#846)
* feat(typesense): sync server state in getOrCreateCollectionFromModel

This commit updates the `getOrCreateCollectionFromModel` method in
`TypesenseEngine` to verify if a collection exists on the server. If not
found, it is created. This ensures consistency between the worker and
server states.

* style: fix styleCI errors

refactor: remove uneeded cast of flag

The collectionExists flag was already cast as false, so recasting it
serves no purpose

* Update TypesenseEngine.php

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>